requirements-dev.txt
Pillow==5.1.0
iterm2-tools
ipdb
flake8
requests
